# openapi.yaml
openapi: 3.0.0
paths:
...
  /api/ponds/{pondId}/telemetry:
    get:
      operationId: TelemetryController_query
      parameters:
        - name: pondId
          required: true
          in: path
          schema:
            type: string
        - name: limit
          required: false
          in: query
          schema:
            type: number
        - name: window
          required: false
          in: query
          schema:
            enum:
              - 15m
              - 1h
              - 6h
              - 24h
              - 7d
              - 30d
            type: string
      responses:
        "200":
          description: ""
      tags: &a7
        - telemetry
      security: &a8
        - bearer: []
  /api/ponds/{pondId}/telemetry/series:
    get:
      operationId: TelemetryController_series
      parameters:
        - name: pondId
          required: true
          in: path
          schema:
            type: string
        - name: metric
          required: true
          in: query
          schema:
            type: string
        - name: since
          required: false
          in: query
          schema:
            enum:
              - 24h
              - 7d
              - 30d
            type: string
      responses:
        "200":
          description: ""
      tags: *a7
      security: *a8
...
  /api/ponds/{pondId}/alerts:
    get:
      operationId: AlertEventsController_list
      parameters:
        - name: pondId
          required: true
          in: path
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: 1-200, default 50
          schema:
            type: number
        - name: onlyOpen
          required: false
          in: query
          schema:
            type: boolean
        - name: severity
          required: false
          in: query
          schema:
            enum:
              - INFO
              - WARNING
              - CRITICAL
            type: string
        - name: since
          required: false
          in: query
          schema:
            enum:
              - 24h
              - 7d
              - 30d
            type: string
      responses:
        "200":
          description: ""
      tags:
        - alert-events
      security:
        - bearer: []
...
  /api/ponds/{pondId}/predictions:
    get:
      operationId: PredictionsController_get
      parameters:
        - name: pondId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: &a11
        - predictions
      security: &a12
        - bearer: []
  /api/ponds/{pondId}/predictions/history:
    get:
      operationId: PredictionsController_history
      parameters:
        - name: pondId
          required: true
          in: path
          schema:
            type: string
        - name: limit
          required: false
          in: query
          schema:
            type: number
        - name: offset
          required: false
          in: query
          schema:
            type: number
        - name: kind
          required: false
          in: query
          schema:
            enum:
              - ANOMALY_SCORE
              - DO_FORECAST
              - NH3_FORECAST
            type: string
        - name: flagged
          required: false
          in: query
          schema:
            enum:
              - "true"
              - "false"
            type: string
        - name: since
          required: false
          in: query
          schema:
            enum:
              - 24h
              - 7d
              - 30d
            type: string
      responses:
        "200":
          description: ""
      tags: *a11
      security: *a12
...
info:
  title: Koi Pond Water Quality Monitor
  description: Server API for pond telemetry, alerts, and recommendations
  version: 0.1.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
...